Adds moderator role and implements authorization rules
https://www.bloc.io/users/danya-smith/checkpoints/1644?roadmap_section=120

In this checkpoint, we learn that we can add a Role field to the database
and use the enum structure to populate it with roles. In this assignment,
we add the :moderator role to existing roles, and provide it with the
ability to edit Topics, and edit and create Posts.

+++ app/controllers/topics_controller.rb
@@ -3,6 +3,7 @@ class TopicsController < ApplicationController
   before_action :require_sign_in, except: [:index, :show]
   before_action :authorize_user, except: [:index, :show]
 
+
   def index
     @topics = Topic.all
   end
@@ -12,16 +13,26 @@ class TopicsController < ApplicationController
   end
 
   def new
-    @topic = Topic.new
+    if current_user.admin?
+      @topic = Topic.new
+    else
+      flash[:error] = "Moderators cannot create topics."
+      redirect_to action: :index
+    end
   end
 
   def create
-    @topic = Topic.new(topic_params)
-    if @topic.save
-      redirect_to @topic, notice: "Topic was saved successfully."
+    if current_user.admin?
+      @topic = Topic.new(topic_params)
+      if @topic.save
+        redirect_to @topic, notice: "Topic was saved successfully."
+      else
+        flash[:error] = "Error creating topic. Please try again."
+        render :new
+      end
     else
-      flash[:error] = "Error creating topic. Please try again."
-      render :new
+      flash[:error] = "Moderators cannot create topics."
+      redirect_to action: :index
     end
   end
 
@@ -46,15 +57,21 @@ class TopicsController < ApplicationController
   end
 
   def destroy
-    @topic = Topic.find(params[:id])
+    if current_user.admin?
+      @topic = Topic.find(params[:id])
 
-    if @topic.destroy
-      flash[:notice] = "\"#{@topic.name}\" was deleted successfully."
-      redirect_to action: :index
+      if @topic.destroy
+        flash[:notice] = "\"#{@topic.name}\" was deleted successfully."
+        redirect_to action: :index
+      else
+        flash[:error] = "There was an error deleting the topic."
+        render :show
+      end
     else
-      flash[:error] = "There was an error deleting the topic."
-      render :show
+      flash[:error] = "Moderators cannot delete topics."
+      redirect_to action: :index        
     end
+
   end
 
   private
@@ -64,9 +81,15 @@ class TopicsController < ApplicationController
   end
 
   def authorize_user
-    unless current_user.admin?
+
+    if params[:action] == "destroy" && current_user.admin? == false
       flash[:error] = "You must be an admin to do that."
       redirect_to topics_path
-    end
+    else
+      unless current_user.admin? || current_user.moderator?
+        flash[:error] = "You must be an admin to do that."
+        redirect_to topics_path
+      end      
+    end      
   end
 end
